+++ data_juicer/core/ray_executor.py
@@ -1,58 +1,15 @@
-import os
 import time
 
-import pyarrow as pa
 from loguru import logger
 
-from data_juicer import cuda_device_count, use_cuda
 from data_juicer.config import init_configs
-from data_juicer.ops import Filter, Mapper, load_ops
+from data_juicer.core.ray_data import RayDataset
+from data_juicer.ops import load_ops
 from data_juicer.utils.availability_utils import AvailabilityChecking
-from data_juicer.utils.constant import Fields
-from data_juicer.utils.process_utils import calculate_np
 
 with AvailabilityChecking(['ray'], requires_type='dist'):
     import ray
     import ray.data as rd
-
-
-def is_valid_path(item, dataset_dir):
-    full_path = os.path.abspath(os.path.join(dataset_dir, item))
-    return os.path.exists(full_path)
-
-
-def convert_to_absolute_paths(dict_with_paths, dataset_dir, path_keys):
-    for key in path_keys:
-        if key not in dict_with_paths:
-            continue
-        if isinstance(dict_with_paths[key], list):
-            dict_with_paths[key] = [
-                os.path.abspath(os.path.join(dataset_dir, item))
-                if isinstance(item, str) and is_valid_path(dataset_dir, item)
-                else item for item in dict_with_paths[key]
-            ]
-        elif isinstance(dict_with_paths[key], str):
-            dict_with_paths[key] = os.path.abspath(
-                os.path.join(dataset_dir,
-                             dict_with_paths[key])) if is_valid_path(
-                                 dict_with_paths[key],
-                                 dataset_dir) else dict_with_paths[key]
-    return dict_with_paths
-
-
-def set_dataset_to_absolute_path(dataset, dataset_path, cfg):
-    """
-    Set all the path in input data to absolute path.
-    Checks dataset_dir and project_dir for valid paths.
-    """
-    if not (cfg.video_key in dataset.columns() or cfg.image_key
-            in dataset.columns() or cfg.audio_key in dataset.columns()):
-        return dataset
-    dataset_dir = os.path.dirname(dataset_path)
-    dataset = dataset.map(lambda item: convert_to_absolute_paths(
-        item, dataset_dir, [cfg.video_key, cfg.image_key, cfg.audio_key]))
-    logger.info(f"transfer {dataset.count()} sample's paths")
-    return dataset
 
 
 class RayExecutor:
@@ -83,42 +40,6 @@ class RayExecutor:
         ray.init(self.cfg.ray_address)
         self.process_list = self.cfg.process
 
-    def get_num_gpus(self, op, op_proc):
-        if not use_cuda() or not op._accelerator == 'cuda':
-            return 0
-        proc_per_gpu = op_proc / cuda_device_count()
-        return 1.0 / proc_per_gpu
-
-    def run_op(self, op, op_cfg, dataset):
-        op_name = next(iter(op_cfg.keys()))
-        op_proc = calculate_np(op_name, op.mem_required, op.cpu_required,
-                               self.cfg.np)
-        num_gpus = self.get_num_gpus(op, op_proc)
-        try:
-            if isinstance(op, Mapper):
-                dataset = dataset.map_batches(op.process,
-                                              batch_size=1,
-                                              batch_format='pyarrow',
-                                              num_gpus=num_gpus)
-            elif isinstance(op, Filter):
-                dataset = dataset.map_batches(op.compute_stats,
-                                              batch_size=1,
-                                              batch_format='pyarrow',
-                                              num_gpus=num_gpus)
-                if op.stats_export_path is not None:
-                    dataset.write_json(op.stats_export_path, force_ascii=False)
-                dataset = dataset.filter(op.process)
-            else:
-                logger.error(
-                    'Ray executor only support Filter and Mapper OPs for now')
-                raise NotImplementedError
-        except:  # noqa: E722
-            logger.error(f'An error occurred during Op [{op_name}].')
-            import traceback
-            traceback.print_exc()
-            exit(1)
-        return dataset
-
     def run(self, load_data_np=None):
         """
         Running the dataset process pipeline.
@@ -131,37 +52,20 @@ class RayExecutor:
         dataset = rd.read_json(self.cfg.dataset_path)
 
         # convert all the path in dataset to absolute path
-        dataset = set_dataset_to_absolute_path(dataset, self.cfg.dataset_path,
-                                               self.cfg)
-        logger.info('Dataset columns:', dataset.columns())
+        dataset = RayDataset(dataset, self.cfg.dataset_path, self.cfg)
         # 2. extract processes
         logger.info('Preparing process operators...')
         self.process_list, self.ops = load_ops(self.cfg.process,
                                                self.cfg.op_fusion)
 
         # 3. data process
-        # - If tracer is open, trace each op after it's processed
-        # - If checkpoint is open, clean the cache files after each process
-        if Fields.stats not in dataset.columns(fetch_if_missing=False):
-            logger.info(f'columns {dataset.columns(fetch_if_missing=False)}')
-
-            def process_batch_arrow(table: pa.Table) -> pa.Table:
-                new_column_data = [{} for _ in range(len(table))]
-                new_talbe = table.append_column(Fields.stats,
-                                                [new_column_data])
-                return new_talbe
-
-            dataset = dataset.map_batches(process_batch_arrow,
-                                          batch_format='pyarrow')
-
         logger.info('Processing data...')
         tstart = time.time()
-        for op_cfg, op in zip(self.process_list, self.ops):
-            dataset = self.run_op(op, op_cfg, dataset)
+        dataset.process(self.ops)
 
         # 4. data export
         logger.info('Exporting dataset to disk...')
-        dataset.write_json(self.cfg.export_path, force_ascii=False)
+        dataset.data.write_json(self.cfg.export_path, force_ascii=False)
         tend = time.time()
         logger.info(f'All Ops are done in {"%.3f" % (tend - tstart)}(s).')
         return dataset
