+++ tests/run.py
@@ -54,6 +54,7 @@ def gather_test_cases(test_dir, pattern, tag):
           f'{SKIPPED_TESTS.modules}')
     for suite_discovered in discover:
         for test_suite in suite_discovered:
+            print("Discover", test_suite)
             for test_case in test_suite:
                 if type(test_case) in SKIPPED_TESTS.modules.values():
                     continue
